Fix realtime address balance update
The balance update stopped working because the client was trying to
subscribe twice in the channel topic. When the server receives a
duplicated subscription, he closes the first connection to open a new
one.

In our case, the client subscribed in the address channel in address.js
and again address/transactions.js, then the server was keeping only the
last subscription.